where_to_go/requirements.txt
asgiref==3.2.10
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
contextlib2==21.6.0
crashtest==0.3.1
cryptography==37.0.4
distlib==0.3.6
Django==3.1
django-admin-sortable2==1.0.4
django-tinymce==3.4.0
docopt==0.6.2
dulwich==0.20.46
environs==9.5.0
filelock==3.8.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
marshmallow==3.19.0
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
Pillow==9.3.0
pipenv==2022.11.11
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pythonanywhere==0.10.3
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.9.1
schema==0.7.5
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tomlkit==0.11.5
typer==0.7.0
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==6.3.0
